restore `fillchars` even if it was empty
`s:toggle_minimal_look` only restored `fillchars` if it was originally
not empty.  This led to a bug in Neovim.

If `fillchars` is empty, both Vim and Neovim behave as if it was set to
`eob:~`.  Vim's default for `fillchars` is `vert:|,fold:-,eob:~`, so
`s:toggle_minimal_look` restored `eob:~`.  But Neovim's default is
empty, so nothing was restored and the minimal-look setting, `eob: `,
remained in effect after deactivation.